+++ source/isaaclab/isaaclab/sensors/imu/imu_cfg.py
@@ -7,8 +7,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from isaaclab.markers import VisualizationMarkersCfg
-from isaaclab.markers.config import RED_ARROW_X_MARKER_CFG
 from isaaclab.utils import configclass
 
 from ..sensor_base_cfg import SensorBaseCfg
@@ -41,16 +39,3 @@ class ImuCfg(SensorBaseCfg):
 
     offset: OffsetCfg = OffsetCfg()
     """The offset pose of the sensor's frame from the sensor's parent frame. Defaults to identity."""
-
-    visualizer_cfg: VisualizationMarkersCfg = RED_ARROW_X_MARKER_CFG.replace(prim_path="/Visuals/Command/velocity_goal")
-    """The configuration object for the visualization markers. Defaults to RED_ARROW_X_MARKER_CFG.
-
-    This attribute is only used when debug visualization is enabled.
-    """
-    gravity_bias: tuple[float, float, float] = (0.0, 0.0, 9.81)
-    """The linear acceleration bias applied to the linear acceleration in the world frame [m/s^2] (x, y, z).
-
-    IMU sensors typically output a positive gravity acceleration in opposition to the direction of gravity. This
-    config parameter allows users to subtract that bias if set to (0., 0., 0.). By default this is set to (0.0, 0.0,
-    9.81) which results in a positive acceleration reading in the world Z.
-    """